Refactor to wrap everything in a function
